apps/Screenshot_Tool/requirements.txt
PyQt5
Pillow
pyinstaller

